requirements.txt
PyPDF2
python-docx
requests
qdrant-client
sentence-transformers